+++ docs/source/conf.py
@@ -45,3 +45,9 @@ html_theme = 'sphinx_book_theme'
 html_static_path = ['_static']
 html_css_files = ["css/style.css"]
 html_js_files = ["js/custom.js"]
+
+html_theme_options = {
+    'home_page_in_toc': True,
+    'show_navbar_depth': 2,
+    'show_toc_level': 2,
+}